refactor: domain 모듈 (jpa, entity)
