+++ domain/build.gradle
@@ -1,27 +1,17 @@
 plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.3.6'
-    id 'io.spring.dependency-management' version '1.1.6'
+    id 'org.springframework.boot'
 }
 
-group = 'com.example'
-version = '0.0.1-SNAPSHOT'
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(17)
-    }
+bootJar {
+    enabled = false
 }
 
-allprojects {
-    repositories {
-        mavenCentral()  // 모든 프로젝트에서 Maven Central을 사용할 수 있게 설정
-    }
+jar {
+    enabled = true
 }
 
 dependencies {
     implementation project(':core')
-    implementation project(':domain')
 
     implementation 'org.springframework.boot:spring-boot-starter'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
@@ -39,14 +29,4 @@ dependencies {
 
     // Swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
-
-subprojects {
-    apply plugin: 'java'
-    apply plugin: 'org.springframework.boot'
-    apply plugin: 'io.spring.dependency-management'
 }
