+++ domain/src/main/java/com/example/banking/domain/account/type/AccountStatus.java
@@ -1,0 +1,6 @@
+package com.example.banking.domain.account.type;
+
+public enum AccountStatus {
+    ACTIVE,     // 활성
+    DELETED     // 삭제됨
+}
