+++ src/main/java/com/example/demo/DemoApplication.java
@@ -2,11 +2,16 @@ package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EntityScan(basePackages = "com.example.banking.domain")
+@EnableJpaRepositories(basePackages = "com.example.banking.domain")
 @ComponentScan(basePackages = {
     "com.example.banking.core",
+    "com.example.banking.domain",
     "com.example.demo"
 })
 public class DemoApplication {
