--- src/main/java/com/example/demo/account/repository/AccountRepository.java
@@ -1,22 +1,0 @@
-package com.example.demo.account.repository;
-
-import com.example.demo.account.entity.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
-import jakarta.persistence.LockModeType;
-import java.util.Optional;
-
-@Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Optional<Account> findByAccountNumber(String accountNumber);
-
-    boolean existsByAccountNumber(String accountNumber);
-
-    @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @Query("SELECT a FROM Account a WHERE a.accountNumber = :accountNumber")
-    Optional<Account> findByAccountNumberWithLock(@Param("accountNumber") String accountNumber);
-}
