--- src/main/java/com/example/demo/account/type/AccountStatus.java
@@ -1,6 +1,0 @@
-package com.example.demo.account.type;
-
-public enum AccountStatus {
-    ACTIVE,     // 활성
-    DELETED     // 삭제됨
-}
