+++ src/main/resources/application.yml
@@ -11,6 +11,10 @@ spring:
     show-sql: true
     properties:
       hibernate:
+        packages-to-scan: com.example.banking.domain
         format_sql: true
         jdbc:
           time_zone: Asia/Seoul
+
+server:
+  port: 8081
